% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_add_ellipses}
\alias{mm_add_ellipses}
\title{Add ellipses to plot}
\usage{
mm_add_ellipses(elist, cols, alpha)
}
\arguments{
\item{elist}{Output of \code{mm_grp_Ellipses}, or an otherwise defined mesh3d object.}

\item{cols}{a vector of colors to plot each ellipse}

\item{alpha}{optional value to specify transparency 0 = invisible, 1 = opqaue.}
}
\description{
Once mesh3d objects are calculated, add them to the current rgl window
}
\details{
mesh3d ellipses must be clalculated first using \code{mm_grp_Ellipses()} and the output saved to an object. That object is used for \code{elist}.
}
