% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_ArrayData}
\alias{mm_ArrayData}
\title{Array Data}
\usage{
mm_ArrayData(IDs, DAYS, VALUE, MID = NULL, avgLENGTH, avgMID = NULL)
}
\arguments{
\item{IDs}{A vector that contains indiviaul IDs repeated for muliple days of collection.}

\item{DAYS}{A vector that contains information on time, IE Day 1, Day 2, Day 3. Note: this vector should include integers, continuous data might produce unintended results.}

\item{VALUE}{A vector containing the variable sampled.}

\item{MID}{A vector of midpoints to center each individuals profile. These should be unique to each individual and repeated for each observation of DAYS, VALUE, and IDs.}

\item{avgLENGTH}{Integer. Number of days to up/down sample observations to using \code{\link{mm_GetInterval}}.}

\item{avgMID}{If NULL (default) 0 will be centered using mm_interval.}
}
\value{
Returns a 3D array of data to be analyzed with individuals in the 3rd dimension.
}
\description{
Construct a ragged array (containing missing data) of a specified length (up/down sampling individuals to fit).
}
