% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mm_CalcShapespace}
\alias{mm_CalcShapespace}
\title{Define Shapespace of aligned dataset.}
\usage{
mm_CalcShapespace(dat, max_Shapes = 10)
}
\arguments{
\item{dat}{A 3D array of shape data to be analyzed.}

\item{max_Shapes}{The maximum amount of PCs to visualize. Default 10.}
}
\value{
A list containing the results of shape-pca, including vizualizations of shape extrema for each Principal Component.
}
\description{
Conduct PCA of shape data and visualize major shape trends.
}
