% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mm_CheckImputation}
\alias{mm_CheckImputation}
\title{Check Imputation}
\usage{
mm_CheckImputation(A1, A2, ObO = TRUE)
}
\arguments{
\item{A1}{An aligned array, containing missing data (presumably made with  \code{\link{mm_ArrayData}}\verb{$shape_data_wNA}).}

\item{A2}{An aligned and imputed array (presumably made with \code{\link{mm_ArrayData}}\verb{$Shape_data}).}

\item{ObO}{One-by-One. If TRUE (default), individuals will be plotted one at a time, requiring the user to advance/exit the operation. If FALSE, all plots will be generated at once to be browsed/exported from  the Plot History}
}
\description{
Plot Raw (alinged) data along side by side with imputed data.
}
