% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_CheckInt}
\alias{mm_CheckInt}
\title{Check Imputation}
\usage{
mm_CheckInt(A1, A2, ObO = TRUE)
}
\arguments{
\item{A1}{An aligned array, containng missing data (presumably made with \code{\link{mm_ArrayData}})}

\item{A2}{An aligned and imputed array (presumbably made with \code{\link{mm_FillMissing}})}

\item{ObO}{One-by-One. If TRUE (default), individuals will be plotted one at a time, requiring the user to advance/exit the operation. If FALSE, all plots will be generated at once to be browsed/exported from  the Plot History}
}
\description{
Plot Raw (alinged) data along side by side with imputed data.
}
