% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mm_ColorLeaves}
\alias{mm_ColorLeaves}
\title{Color leves of a dendrogram}
\usage{
mm_ColorLeaves(dendro, cols)
}
\arguments{
\item{dendro}{A dendrogram or hclust class object}

\item{cols}{a vector of colors}
}
\description{
Specify color order approriately for a dendrogram
}
\details{
Leaves of a dendrogram will be re-ordered compared to most input classifiers. This function takes the study-ordered colors and correctly applies them to the dendrogram using \code{dendextend}
}
