% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mm_CompModel}
\alias{mm_CompModel}
\title{Compare Model Metrics}
\usage{
mm_CompModel(mv_results, row_labels = NULL, digits = 4)
}
\arguments{
\item{mv_results}{Input mvlm, created by mm_BuildModel (or by using geomorph::procD.lm)}

\item{row_labels}{A character vector to use in output. If NULL (default) labels from the input data will be used.}

\item{digits}{Number of decimal places to round to. Default includes 4 decimal places.}
}
\value{
A list containing the results of the mvlm, visualizations of shape
trends along the regression line, and the model itself.
}
\description{
Compare key figs (Rsq, p-value, etc) across multiple models.
}
