% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mm_coords_to_shape}
\alias{mm_coords_to_shape}
\title{Visualize shape of target coordinates}
\usage{
mm_coords_to_shape(A, PCA, target_coords, target_PCs = c(1, 2))
}
\arguments{
\item{A}{A landmark array used for the pca}

\item{PCA}{output of prcomp. Should contain $transormation}

\item{target_coords}{A single set of X,Y coordinates.}

\item{target_PCs}{Integer identifying which pc to use on the X and Y axis.
Default is c(1,2) for PC1 on x and PC2 on y}
}
\value{
A landmark array representing the hypothetical shape of a given set
of coordinates.
}
\description{
Visualize shape of target coordinates
}
