% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mm_data}
\alias{mm_data}
\title{Sample hormone dataset}
\format{
A matrix with 2015 obs (rows) and 4 variables (columns).
\describe{

\item{ID}{Individual id, each integer represents a different individual. Includes leading 0s.}
\item{CYCLEDAY}{Integer day of cycle. Generally runs from 1 ... 28}
\item{MIDPOINT}{Single value for each individual, repeated along each CYCLEDAY. In this sample, day of ovulation.}
\item{E1G}{Daily measure of hormone, in nanograms per mililiter}

}
}
\usage{
mm_data
}
\description{
Sample dataset classifiers to be paired with sample array. This table contains 60 rows to match the 60 individuals across the third dimension of the array
}
\keyword{datasets}
