% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mm_ellipse}
\alias{mm_ellipse}
\title{Add confidence ellipses to an active scatterplot.}
\usage{
mm_ellipse(
  dat,
  ci = c(67.5, 90, 95, 99),
  linesCol = "black",
  fillCol = "grey",
  smoothness = 20
)
}
\arguments{
\item{dat}{A matrix of data to draw an ellipses around.}

\item{ci}{Percentage of data to capture. Must be one of c(67.5, 90, 95, 99).}

\item{linesCol}{Border color of the shape.}

\item{fillCol}{Fill color of the shape.}

\item{smoothness}{Lower values will look jagged, higher value will make
smoother lines, but may take a long time to plot. Default value is 20.}
}
\value{
No value. Will add an ellipses of a given size to the current plot.
}
\description{
Add confidence ellipses to an active scatterplot.
}
