% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mm_FlattenArray}
\alias{mm_FlattenArray}
\title{Flatten Array}
\usage{
mm_FlattenArray(A, sep = ".")
}
\arguments{
\item{A, }{an array to be flattened}

\item{sep}{Separator to be used for column names}
}
\value{
Returns a flattened array
}
\description{
Convert a 3D array to 2D matrix suitable for PCA, etc. Note, this function is identical to geomorph::two.d.array, reproduced here for convenience.
}
