% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_GeomScale}
\alias{mm_GeomScale}
\title{Geometric Scaling}
\usage{
mm_GeomScale(x)
}
\arguments{
\item{x}{A numeric vector to be scaled. Missing values will produce NA, conduct knn imputation using mm_FillMissing first.}
}
\description{
Calculate the geometric mean of a vector and scale all values by it.
}
\examples{
mm_GeomScale(1:10)
}
