% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_grp_dists}
\alias{mm_grp_dists}
\title{Distance from Centroid}
\usage{
mm_grp_dists(dat, grps, plots = TRUE)
}
\arguments{
\item{dat}{a 2d matrix of data. Presumably PC scores}

\item{grps}{a vector defining group IDs}

\item{plots}{Logical. Should distances be plotted as vusing boxplots? If FALSE, distance calculations are still performed}
}
\description{
Calculate and plot group distance from centroid (grand mean)
}
