% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_grp_Ellipses}
\alias{mm_grp_Ellipses}
\title{Calculate elipses by groups}
\usage{
mm_grp_Ellipses(dat, grps, confidence = 0.95)
}
\arguments{
\item{dat}{A 2D matrix, most likely PC scores}

\item{grps}{A vector of group IDs}

\item{confidence}{CI to draw ellipses at. Default is 95\%}
}
\value{
If successful, returns a mesh3d object (list) to be plotted with rgl functions. See \code{mm_add_ellipses}.
}
\description{
Calculate coordinates of 3D elipsoids using \code{rgl}
}
