% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_MinMaxScale}
\alias{mm_MinMaxScale}
\title{Min-Max Scaling}
\usage{
mm_MinMaxScale(x)
}
\arguments{
\item{x}{A Numeric vector to be scaled. Missing values are allowed and ignored.}
}
\value{
Returns a scaled vector
}
\description{
Scale a vector from 0,1 based on its minimum and maximum values.
}
\examples{
mm_MinMaxScale(1:10)
}
