% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mm_mute_cols}
\alias{mm_mute_cols}
\title{Take a color and modify it}
\usage{
mm_mute_cols(cols, s = NULL, v = NULL, alpha = 0.4)
}
\arguments{
\item{cols}{a vector of colors, eg: "#0066FF"}

\item{s}{Either a single value or a vector of same length as cols specifying a new saturation (range 0-1). colors darken to black (0)}

\item{v}{Either a single value or a vector of same length as cols specifying a new value (range 0-1). colors lighten to white (0)}

\item{alpha}{Either a single value or a vector of same length as cols specifying a transparency value (range 0-1). colors translucent at 0.}
}
\description{
Modify color/transparency using hsv syntax
}
