% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_pheno_plot}
\alias{mm_pheno_plot}
\title{Visualize PC axes}
\usage{
mm_pheno_plot(pheno, xPC = 1, yPC = 2, yr = c(0, 1), title = "")
}
\arguments{
\item{pheno}{Output of \code{mm_phenotype}, containing a PCA object with PC shapes}

\item{xPC}{The PC to be plotted on the x axis}

\item{yPC}{The PC to be plotted on the y axis}

\item{yr}{The y-xis range, in the format c(0,1)}

\item{title}{To be used for the plot}
}
\description{
Plot a scatterplot and vizualize shape change across the X axis.
}
\details{
Meant to be a quick diagnostic plot with minimal customization.
}
