% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mm_Phenotype}
\alias{mm_Phenotype}
\title{Generate Phenotypes}
\usage{
mm_Phenotype(dat, kgrps, cuttree_h = NULL, cuttree_k = NULL, plot_figs = TRUE)
}
\arguments{
\item{dat}{Either an Array of shape data, an mmPCA object, or an mmDiag object.}

\item{kgrps}{A non-negative integer of sub-groups to draw. kgrps=1 will provide results for the whole input dat.}

\item{cuttree_h}{Optional. Draw clusters by splitting the tree at a given height, h.}

\item{cuttree_k}{Optional. Draw clsuters by splitting the tree into number of branches, k}

\item{plot_figs}{Optional. Default = TRUE, plot phenotypes for each set(s) of subgroups.}
}
\description{
Partition sample into clusters, based on information from
}
