% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mm_pretty_pca}
\alias{mm_pretty_pca}
\title{Plot Calendar Days}
\usage{
mm_pretty_pca(PCA, xPC = 1, yPC = 2, clas_col = NULL, legend_cex = 0.8)
}
\arguments{
\item{PCA}{Input data either prcomp or mmPCA.}

\item{xPC}{The PC to plot on the x axis}

\item{yPC}{The PC to plot on the y axis}

\item{clas_col}{A character vector of groupings. Each level will be plotted as a different color.}

\item{legend_cex}{A scaling factor to be applied specifically to the legend. Set to NULL for scatterplot only.}
}
\value{
Returns no object, plots results of PCA
}
\description{
Pretty PCA
}
\details{
A better PCA plot
}
