% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_ScreePlot}
\alias{mm_ScreePlot}
\title{Scree Plot}
\usage{
mm_ScreePlot(x, maxC = 15, ...)
}
\arguments{
\item{x}{Input data for cluster analysis (IE, PCA)}

\item{maxC}{Maximum clusters to evaluate}

\item{...}{Additional arguments to be passed to plot}
}
\description{
Plot total within group sum of squares to evalaute clusters
}
