% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_SilPlot}
\alias{mm_SilPlot}
\title{Silouhete Width Plot}
\usage{
mm_SilPlot(x, maxC = 15, ...)
}
\description{
Plot average silhouete widths to evaluate clusters
@param x Input data for cluster analysis (IE PCA)
@param maxC Maximum clusters to evaluate
@param ... additional arguments passed to plot
}
\details{
@export
}
