% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_transf_log}
\alias{mm_transf_log}
\title{natural log transform}
\usage{
mm_transf_log(x)
}
\arguments{
\item{x}{A numeric vector to be scaled. Missing values will produce NA, conduct knn imputation using mm_FillMissing first.}
}
\value{
Returns a scaled vector
}
\description{
Transform a vector by the natural log.
}
\examples{
mm_transf_log(1:10)
}
