% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_transf_log10}
\alias{mm_transf_log10}
\title{Common log transform}
\usage{
mm_transf_log10(x)
}
\arguments{
\item{x}{A numeric vector to be scaled. Missing values will produce NA, conduct knn imputation using mm_FillMissing first.}
}
\value{
Returns a scaled vector
}
\description{
Transform a vector by the common log (base 10).
}
\examples{
mm_transf_log10(1:10)
}
