% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_transf_zscore}
\alias{mm_transf_zscore}
\title{Z scores}
\usage{
mm_transf_zscore(x)
}
\arguments{
\item{x}{A numeric vector to be scaled. Missing values will produce NA, conduct knn imputation using mm_FillMissing first.}
}
\value{
Returns a scaled vector
}
\description{
Calculate and return z-scores given a numeric vector.
}
\examples{
mm_transf_zscore(1:10)
}
