% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mm_VizModel}
\alias{mm_VizModel}
\title{Visualize Multivariate LM}
\usage{
mm_VizModel(dat, clas_col = NULL)
}
\arguments{
\item{dat}{Input mvlm, created by mm_BuildModel (or by using geomorph::procD.lm)}

\item{clas_col}{A classifier to color the data by. If null (default) all points will be grey. Otherwise, data will be plotted as rainbow(n) colors.}
}
\description{
Visualize 2D scatterplot of mvlm including predicted shapes.
}
