% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_fix.R
\name{new_array}
\alias{new_array}
\title{new alignment function}
\usage{
new_array(ObsIDs, ObsDays, ObsValue, ObsFixed, tar_array, tar_fixed, tar_scl)
}
\arguments{
\item{ObsIDs}{A vector that contains indiviaul IDs repeated for muliple days of collection.}

\item{ObsDays}{A vector that contains information on time, IE Day 1, Day 2, Day 3. Note: this vector should include integers, continuous data might produce unintended results.}

\item{ObsValue}{A vector containing the variable sampled.}

\item{ObsFixed}{A vector containing the days to be treated as fixed. Note: ObsMid must have the same number of observations as unique Individuals.}

\item{tar_array}{A standard target to align individuals. \strong{EXPLAIN}}

\item{tar_fixed}{A sequence identifying fixed landmarks in the target}

\item{tar_scl}{A sequence identifying the scale}
}
\value{
Returns a 3D array of data to be analyzed with individuals in the 3rd dimension.
}
\description{
hopefully more flexible
}
