% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pheno_plot}
\alias{pheno_plot}
\title{Visualize PC axes}
\usage{
pheno_plot(pheno, xPC = 1, yPC = 2, yr = c(0, 1), title = "")
}
\arguments{
\item{pheno}{Output of \code{mm_phenotype}, containing a PCA object with PC shapes}
}
\description{
Plot a scatterplot and vizualize shape change across the X axis.
}
\details{
Meant to be a quick diagnostic plot with minimal customization.
}
