% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mm_Diagnostics}
\alias{mm_Diagnostics}
\title{Run a suite of diagnostic analysis/plots}
\usage{
mm_Diagnostics(dat, max_PC_viz = 10, max_PC_calc = NULL, hide_plots = FALSE)
}
\arguments{
\item{dat}{A 3D array or a mmPCA object (output of mm_CalcShapespace).}

\item{max_PC_viz}{Maximum number of PCs to include in visualizations (EG Eigenplots, or shape trends.}

\item{max_PC_calc}{By default (NULL), all PCs will be included in calculations. However, if fewer PCs are required users may specify an integer, n, to get the first n PCS.}

\item{hide_plots}{By default (FALSE), helpful visuals are plotted.}
}
\description{
Run a suite of diagnostic analysis/plots
}
