% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_FillMissing}
\alias{mm_FillMissing}
\title{Impute Missing Data}
\usage{
mm_FillMissing(A, knn = 3, scale = c("none", "MinMax", "Geom"))
}
\arguments{
\item{A}{A ragged array, presumably constructed with \code{\link{mm_ArrayData}}.}

\item{knn}{Number of nearest neighbors to draw on for imputation (default = 3).}

\item{scale}{Type of scaling to implement (or not). Must be one of "none", "MinMax", "Geom"}
}
\description{
Fill in a ragged away by nearest neighbor imputation
}
