% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{mm_FillMissing}
\alias{mm_FillMissing}
\title{Impute Missing Data}
\usage{
mm_FillMissing(A, knn = 3)
}
\arguments{
\item{A}{A ragged array (IE, contains missing cells), presumably constructed with \code{\link{mm_ArrayData}}.}

\item{knn}{Number of nearest neighbors to draw on for imputation (default = 3).}
}
\description{
Fill in a ragged away by nearest neighbor imputation
}
