% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mm_Phenotype}
\alias{mm_Phenotype}
\title{Get Phenotypes}
\usage{
mm_Phenotype(A, k = NULL, maxPC = 10)
}
\arguments{
\item{A}{an array to be analyzed. No missing data allowed (see \code{\link{mm_FillMissing}})}

\item{k}{The number of groups for which phenotypes will be drawn. If NULL (default) Diagnostic plots will be drawn to help gauge approprite k}

\item{maxPC}{Maximum number of PCs to use in analysis.}
}
\description{
Apply PCA and Hierarchical Clustering to infer phenotypes
}
