% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mm_SilPlot}
\alias{mm_SilPlot}
\title{Silhouette Width Plot}
\usage{
mm_SilPlot(x, maxC = 15, ...)
}
\arguments{
\item{x}{Input data for cluster analysis (IE PCA)}

\item{maxC}{Maximum clusters to evaluate}

\item{...}{additional arguments passed to plot}
}
\description{
Plot average silhouete widths to evaluate clusters
}
