% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moRphomenses_PACKAGE.R
\name{moRphomenses}
\alias{moRphomenses}
\alias{moRphomenses-package}
\title{Geometric Morphometric Analysis of Hormone Cycle Phenotypes}
\description{
Analyze shapes/phenotypes of hormone data using Geometric Morphometric inspired methods.
}
\author{
Daniel E. Ehrlich
}
