% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{print_summary}
\alias{print_summary}
\title{Print basic summary}
\usage{
print_summary(aln, grps = NULL)
}
\arguments{
\item{aln}{An object created with mm_ArrayData}

\item{grps}{(Optional) A numeric vector that defines groupings}
}
\description{
Print basic summary
}
